# content/contracts/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PulseGlucoseApi",
    "version": "1.0.0",
    "description": "Machine readable OpenAPI contract for PulseGlucoseApi consumer and operational endpoints.",
    "x-generatedAt": "2026-03-16T04:45:55.431Z"
  },
  "servers": [
    {
      "url": "https://api.your-domain.com"
    }
  ],
  "paths": {
...
    "/api/v1/notifications/stream": {
      "get": {
        "operationId": "notifications-stream",
        "summary": "Notification stream (SSE)",
        "description": "Realtime notification feed with optional sequence cursor.",
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "afterSeq",
            "in": "query",
            "required": false,
            "description": "Start streaming events after this sequence number.",
            "schema": {
              "type": "string",
              "example": "42"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "text/event-stream": {
                "example": {
                  "event": "notification",
                  "data": {
                    "id": "8b6c9a89-4d5f-4fd6-9e21-5d94b68c5b62",
                    "seq": 43,
                    "type": "settings.profile.updated",
                    "title": "Shared settings updated",
                    "message": "Profile settings were updated",
                    "payload": {
                      "updatedAt": "2026-02-16T12:00:00.000Z"
                    },
                    "timestamp": "2026-02-16T12:00:00.000Z",
                    "sender": {
                      "apiKeyId": null,
                      "apiKeyName": null
                    }
                  }
                }
              }
            }
          }
        },
        "x-pulseglucose-auth": "api_key",
        "x-pulseglucose-source": "mixed"
      }
    },
...
    "/api/v1/stream/glucose": {
      "get": {
        "operationId": "glucose-stream",
        "summary": "Glucose stream (SSE)",
        "description": "Realtime glucose updates from official, share, or both sources.",
        "tags": [
          "Streams"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": false,
            "description": "official, share, or both. Defaults to both.",
            "schema": {
              "type": "string",
              "example": "share"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "text/event-stream": {
                "example": {
                  "event": "glucose_update",
                  "data": {
                    "source": "share_api",
                    "reading": {
                      "valueMmolL": 7.4,
                      "valueMgDl": 133,
                      "unit": "mmol/L",
                      "trend": "flat",
                      "trendRate": 0,
                      "trendNumber": 4,
                      "timestamp": "2026-02-16T12:00:00.000Z",
                      "status": "ok",
                      "source": "share_api"
                    }
                  }
                }
              }
            }
          }
        },
        "x-pulseglucose-auth": "api_key",
        "x-pulseglucose-source": "official"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "AdminBearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "StatusTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-status-token"
      }
    }
  }
}